requirements.txt
flask==1.1.1
flask-cors==3.0.8
flask-api==2.0
pyjwt
google-cloud-datastore
